feat: feat added Throwable inside Err
this allows better ease of debugging when we want to
throw an exception upon the encounter of an error,
having the stackTrace inside the Err allows for faster
debugging when throwing from a place in the code.

This concept is very similar to the `with_context`
of the Rust library `anyhow`, which embeds a stackTrace
inside the error variant.